pyproject.toml
[project]
name = "vz-aggregator"
version = "0.1.0"
description = "MVP datového agregátoru veřejných zakázek pro CZ/EU"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.32",
    "beautifulsoup4>=4.12",
    "lxml>=5.2",
    "tenacity>=8.4",
    "loguru>=0.7",
    "pydantic>=2.7",
    "psycopg[binary]>=3.2",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "python-dotenv>=1.0"
]

[project.optional-dependencies]
dev = [
    "black>=24.0",
    "ruff>=0.4",
    "pytest>=8.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "UP", "B", "N", "W"]
ignore = ["E501"]

[tool.ruff.isort]
known-first-party = ["core", "adapters"]
